chore(scripts): split downlevel-dts to multiple files
