+++ scripts/downlevel-dts/getAllFiles.mjs
@@ -1,0 +1,16 @@
+import { readdirSync, statSync } from "fs";
+
+export const getAllFiles = (dirPath, arrayOfFiles = []) => {
+  const files = readdirSync(dirPath);
+
+  files.forEach((file) => {
+    const { isDirectory } = statSync(dirPath + "/" + file);
+    if (isDirectory()) {
+      arrayOfFiles = getAllFiles(dirPath + "/" + file, arrayOfFiles);
+    } else {
+      arrayOfFiles.push(join(dirPath, "/", file));
+    }
+  });
+
+  return arrayOfFiles;
+};
