+++ scripts/downlevel-dts/index.mjs
@@ -1,11 +1,10 @@
 // @ts-check
 import parallelLimit from "async/parallelLimit";
-import { readdirSync, readFileSync, statSync } from "fs";
 import { cpus } from "os";
-import { join } from "path";
 import yargs from "yargs";
 
 import { downlevelWorkspace } from "./downlevelWorkspace.mjs";
+import { getWorkspaces } from "./getWorkspaces.mjs";
 
 // ToDo: Write downlevel-dts as a yargs command, and import yargs in scripts instead.
 yargs
@@ -17,30 +16,7 @@ yargs
   .help()
   .alias("h", "help").argv;
 
-const { packages } = JSON.parse(readFileSync(join(process.cwd(), "package.json")).toString()).workspaces;
-
-const getAllFiles = (dirPath, arrayOfFiles = []) => {
-  const files = readdirSync(dirPath);
-
-  files.forEach((file) => {
-    if (statSync(dirPath + "/" + file).isDirectory()) {
-      arrayOfFiles = getAllFiles(dirPath + "/" + file, arrayOfFiles);
-    } else {
-      arrayOfFiles.push(join(dirPath, "/", file));
-    }
-  });
-
-  return arrayOfFiles;
-};
-
-const workspaces = packages
-  .map((dir) => dir.replace("/*", ""))
-  .flatMap((workspacesDir) =>
-    readdirSync(join(process.cwd(), workspacesDir), { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .map((dirent) => ({ workspacesDir, workspaceName: dirent.name }))
-  );
-
+const workspaces = getWorkspaces(process.cwd());
 const tasks = workspaces.map(({ workspacesDir, workspaceName }) => async () => {
   await downlevelWorkspace(workspacesDir, workspaceName);
 });
